Converting generators to RDFa in VO styling
These changes reuse more existing components than the previous
editions.  It also ensures that there's RDFa inside of the responses,
helping us share our knowledge with the robots of our world.

There are quite a few unused generated files at this stage.  We're
accepting these for now and intend to clean up and extend this sort of
generator in the near future.  The further additions will help make
the application more flexible.  There are further options in terms of
making the application more reusable with this broad space, these can
also be tackled in the future.